pyproject.toml
[project]
name = "bdx-r-mjlab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Louis Le Lay", email = "louislelay@pal-robotics.com" }
]
requires-python = ">=3.13"
dependencies = [
    "mjlab>=0.0.4",
    "mujoco-warp",
]

[project.scripts]
bdx-r-mjlab = "bdx_r_mjlab:main"

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
mujoco-warp = { git = "https://github.com/google-deepmind/mujoco_warp" }
